+++ plom_server/BuildPaperPDF/services/__init__.py
@@ -2,5 +2,8 @@
 # Copyright (C) 2022 Edith Coates
 # Copyright (C) 2022 Brennen Chiu
 # Copyright (C) 2023 Andrew Rechnitzer
+# Copyright (C) 2025 Colin B. Macdonald
+
+"""Services of the Plom Server BuildPaperPDF app."""
 
 from .build_papers import huey_build_single_paper, BuildPapersService