+++ plom_server/Papers/services/__init__.py
@@ -1,7 +1,9 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (C) 2022-2025 Andrew Rechnitzer
 # Copyright (C) 2022-2023 Edith Coates
-# Copyright (C) 2023 Colin B. Macdonald
+# Copyright (C) 2023, 2025 Colin B. Macdonald
+
+"""Services of the Plom Server Paper app."""
 
 from .paper_creator import PaperCreatorService
 from .paper_info import PaperInfoService, fixedpage_version_count