+++ plom_server/Rubrics/services/rubric_permissions.py
@@ -6,7 +6,7 @@ import math
 from copy import deepcopy
 from typing import Any
 
-from plom_server.Base.models import SettingsModel
+from plom_server.Base.services import Settings
 from .utils import pin_to_fractional_nth
 
 
@@ -89,7 +89,11 @@ class RubricPermissionsService:
         rubric_fractional_options = deepcopy(_frac_opt_table)
         # figure out which are currently checked by checking settings
         for opt in rubric_fractional_options:
-            opt["checked"] = SettingsModel.cget(opt["name"])
+            name = opt["name"]
+            assert isinstance(name, str)
+            opt["checked"] = (
+                True if Settings.key_value_store_get_or_none(name) else False
+            )
         rubric_fractional_options = [
             opt
             for opt in rubric_fractional_options
@@ -110,16 +114,16 @@ class RubricPermissionsService:
         for opt in _frac_opt_table:
             a = str(opt["name"])
             if rp.get(a) == "on":
-                SettingsModel.cset(a, True)
+                Settings.key_value_store_set(a, True)
             else:
-                SettingsModel.cset(a, False)
+                Settings.key_value_store_set(a, False)
         for opt in _frac_opt_table:
             a = str(opt["name"])
             implies = opt["implies"]
             assert isinstance(implies, list)  # help mypy
-            if SettingsModel.cget(a):
+            if Settings.key_value_store_get_or_none(a):
                 for i in implies:
-                    SettingsModel.cset(i, True)
+                    Settings.key_value_store_set(i, True)
 
     @staticmethod
     def pin_to_allowed_fraction(v: float | int) -> float:
@@ -137,16 +141,17 @@ class RubricPermissionsService:
             # zero fractional part, nothing to do here
             return v
 
-        s = SettingsModel.load()
         for opt in _frac_opt_table:
             name = opt["name"]
+            assert isinstance(name, str)
             N = opt["denom"]
             assert isinstance(N, int)  # help mypy
             readable_denom = opt["readable"]
             vpin = pin_to_fractional_nth(v, N)
             if vpin is None:
                 continue
-            if not s.get(name):
+            # TODO: query them all at once for better DB access?
+            if not Settings.key_value_store_get_or_none(name):
                 raise ValueError(
                     f"{readable_denom}-point rubrics are currently not allowed"
                 )