+++ plom_server/Scan/views/__init__.py
@@ -2,8 +2,11 @@
 # Copyright (C) 2022 Edith Coates
 # Copyright (C) 2022-2023 Brennen Chiu
 # Copyright (C) 2023-2025 Andrew Rechnitzer
-# Copyright (C) 2024 Colin B. Macdonald
+# Copyright (C) 2024-2025 Colin B. Macdonald
 # Copyright (C) 2025 Deep Shah
+# Copyright (C) 2025 Aidan Murphy
+
+"""Views of the Plom Server Scan app."""
 
 from .scanner_home import (
     ScannerOverview,
@@ -27,7 +30,9 @@ from .pushed_images import (
 
 from .manage_bundle import (
     GetBundlePageFragmentView,
+    ThumbnailContainerFragmentView,
     BundleThumbnailsView,
+    BundleThumbnailsSummaryFragmentView,
     GetBundleThumbnailView,
     BundleLockView,
     BundlePushCollisionView,
@@ -56,10 +61,4 @@ from .cast_image_state import (
     UnknowifyAllDiscardsHTMXView,
 )
 
-from .rotate_images import (
-    RotateImageOneEighty,
-    RotateImageClockwise,
-    RotateImageCounterClockwise,
-    GetRotatedBundleImageView,
-    GetRotatedPushedImageView,
-)
+from .rotate_images import RotateImageView